moved shared types to a new shared folder
